2025ver/web_version/requirements.txt
flask==2.2.3
mysql-connector-python==8.0.32
pandas==1.5.3
python-dateutil==2.8.2
Werkzeug==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
click==8.1.3